requirements.txt
django==5.1.15
djangorestframework==3.17.1
pillow==12.2.0
psycopg2-binary==2.9.11
weasyprint==68.1
